requirements.txt
networkx
tensorboardX
torch
torchsummary
numpy
mpl-toolkits.clifford
matplotlib
imageio